requirements.txt
amqp==2.5.2
backcall==0.1.0
billiard==3.6.2.0
celery==4.4.0
decorator==4.4.1
Django==2.2.1
django-debug-toolbar==2.1
entrypoints==0.3
Faker==3.0.0
flake8==3.7.9
importlib-metadata==1.5.0
ipython==7.11.0
ipython-genutils==0.2.0
jedi==0.15.2
kombu==4.6.7
mccabe==0.6.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
zipp==2.1.0
